Refactor database classes into separate vault_database module
Move VaultDatabaseConnection, VaultDatabaseStaticRoles, VaultDatabaseDynamicRoles, and Database classes from vault_client.py to new vault_database.py module. This reduces vault_client.py from 1052 to 622 lines (~40% reduction) and improves code organization.

All modules and tests updated to import from vault_database. Since this is being deployed to main and backported together, no backward compatibility re-exports are needed.
