+++ changelogs/fragments/81-database-static-roles.yaml
@@ -1,11 +1,11 @@
 ---
 minor_changes:
-  - module_utils/vault_client.py - Add ``VaultDatabaseStaticRoles`` class
+  - module_utils/vault_database.py - Add ``VaultDatabaseStaticRoles`` class
     for managing database static roles with ``list_static_roles()``,
     ``read_static_role()``, ``create_or_update_static_role()``,
     ``delete_static_role()``, and ``get_static_role_credentials()``
     (https://github.com/ansible-collections/hashicorp.vault/pull/81).
-  - module_utils/vault_client.py - Add ``Database`` container class to
+  - module_utils/vault_database.py - Add ``Database`` container class to
     organize database-related clients (connections and static_roles) for
     a specific mount path
     (https://github.com/ansible-collections/hashicorp.vault/pull/81).
