+++ changelogs/fragments/91-database-dynamic-roles.yaml
@@ -1,10 +1,10 @@
 ---
 minor_changes:
-  - module_utils/vault_client.py - Add ``VaultDatabaseDynamicRoles`` class
+  - module_utils/vault_database.py - Add ``VaultDatabaseDynamicRoles`` class
     for managing database dynamic roles with ``list_dynamic_roles()``,
     ``read_dynamic_role()``, ``create_or_update_dynamic_role()``, and
     ``delete_dynamic_role()``
     (https://github.com/ansible-collections/hashicorp.vault/pull/91).
-  - module_utils/vault_client.py - Update ``Database`` container class to
+  - module_utils/vault_database.py - Update ``Database`` container class to
     include ``dynamic_roles`` client for managing dynamic database roles
     (https://github.com/ansible-collections/hashicorp.vault/pull/91).
