+++ tests/unit/plugins/module_utils/test_vault_database_connection.py
@@ -12,6 +12,8 @@ import pytest
 
 from ansible_collections.hashicorp.vault.plugins.module_utils.vault_client import (
     VaultClient,
+)
+from ansible_collections.hashicorp.vault.plugins.module_utils.vault_database import (
     VaultDatabaseConnection,
 )
 from ansible_collections.hashicorp.vault.plugins.module_utils.vault_exceptions import (
